Moved RestaurantItem to restaurant package
