+++ app/src/main/java/com/example/deliveryapp/ui/screens/home/HomeScreen.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.deliveryapp.model.Restaurant
-import com.example.deliveryapp.ui.RestaurantItem
+import com.example.deliveryapp.ui.components.restaurant.RestaurantItem
 import com.example.deliveryapp.ui.components.shared.AppNavigationBar
 import kotlinx.serialization.Serializable
